updates demo app
