--- demo/nextjs-ssr-app/app/App.tsx
@@ -1,11 +1,0 @@
-import Main from "../components/Main";
-
-function App() {
-  return (
-    <div className="container">
-      <Main />
-    </div>
-  );
-}
-
-export default App;
