+++ demo/nextjs-ssr-app/app/layout.tsx
@@ -2,6 +2,8 @@ import "./globals.css";
 
 import { Inter } from "next/font/google";
 
+import Provider from "@/components/provider";
+
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata = {
@@ -13,7 +15,9 @@ export const metadata = {
 export default function RootLayout({ children }: { children: React.ReactNode }) {
   return (
     <html lang="en">
-      <body className={inter.className}>{children}</body>
+      <body className={inter.className}>
+        <Provider>{children}</Provider>
+      </body>
     </html>
   );
 }
