+++ demo/nextjs-ssr-app/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@tanstack/react-query": "^5.37.1",
     "@web3auth/modal": "file:../../packages/modal",
-    "next": "13.4.9",
+    "next": "14.1.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "wagmi": "^2.14.16"
